+++ source/Meadow.Core/Hardware/PwmPin.cs
@@ -7,8 +7,8 @@ namespace Meadow.Hardware
     /// </summary>
     public class PwmPin : PwmPinBase
     {
-        public PwmPin (string name, object key, double minimumFrequency = 0,
-                       double maximumFrequency = 100000)
+        public PwmPin (string name, object key, float minimumFrequency = 0,
+                       float maximumFrequency = 100000)
             : base(name, key, minimumFrequency, maximumFrequency)
         {
